FitSpec

Description

FitSpec tries to find minimal and complete subsets of a specification via black-box mutation testing.

NOTE: * The API is likely to change in the near future.

Example -- properties over not:

import Test.Check
import MinProps

propertyMap :: Int -> (Bool -> Bool) -> [Bool]
propertyMap not n =
  [ holds n $ \p -> not (not p) == p
  , holds n $ \p -> not p /= p
  ,                 not True == False
  ]

main = report not pMap

FitSpec should report that the minimal (equivalent) specifications are either:

2

1 and 3

Documentation

data Args a

Extra arguments / configuration for reportWith. See args for default values.

Constructors

Args

Fields

nMutants :: Int

(starting) number of black-box mutations

minimumTime :: Int

minimum time to run, use 0 for just nMutants

nTestsF :: Int -> Int

number of tests in function of number of mutants

callNames :: [String]

function call templates: ["foo x y","goo x y"]

limitResults :: Maybe Int

Just a limit for results, Nothing for all

extraMutants :: [a]

extra mutants to try to kill alongside mutations

showPropertySets :: [String] -> String

function to show property sets.

showMoreEI :: Bool

show more equivalences and implications

showMutantN :: [String] -> a -> a -> String

special mutant show

args :: ShowMutable a => Args a

Default arguments for reportWith:

nMutants = 500, start with 500 mutants

minimumTime = 5@, keep incresing the number of mutants until 5 seconds elapse

nTestsF = (*2), use 2 tests more than mutants As a rule of thumb, the number of tests should be proportional to the number of mutants. There is *no* general rule of thumb for the exact proportion: in some applications, less is better, in some applications, more is better. Increase this if you spot a false positive. Other good values for this might be:

(*100)        -- more tests, less false positives
(`div` 100)   -- less tests, less false negatives
(const 1000)  -- specific number of tests

callNames = [], use internal default function call template:

["f x y z w x' y' z' ...","g ...","h ...","f' ...",...]

limitResults = Just 3, limit to just 3 results

extraMutants = [], no extra mutants

showPropertySets = unwords, just join property-sets by spaces. Other good values for this might be:

unlines            -- one per line
unwords . take 5   -- separated by spaces, limit to 5
unlines . take 5   -- one per line, limit to 5
take 30 . unwords  -- limit to 30 characters

fixargs :: ShowMutable a => Int -> Int -> Args a

report :: (ShowMutable a, Mutable a) => a -> (Int -> a -> [Bool]) -> IO ()

Report minimality and completeness results. Uses standard configuration (see args). Needs a function to be mutated and a property map.

reportWith :: Mutable a => Args a -> a -> (Int -> a -> [Bool]) -> IO ()

Same as report but can be configured via 'Args'/'args'.

getRawResults :: Mutable a => [a] -> Int -> a -> (a -> [Bool]) -> [([[Int]], [Maybe a])]

relevantImplications :: Eq i => [[i]] -> [i]

relevantPropertySets :: Eq i => [[i]] -> [[i]]

class ShowMutable a

Minimal complete definition

mutantS