FitSpecC

Documentation

data CArgs

Extra arguments for getResultsWith and reportWith

Constructors

CArgs

Fields

functionNames :: [String]

variableNames :: [String]

nResults :: Maybe Int

cargs :: CArgs

Default arguments

getResults1 :: (Eq a, Eq b, Listable a, Listable b) => Int -> (a -> b) -> ((a -> b) -> [Bool]) -> [([Int], Int, Maybe [(a, b)], Int, Maybe [(a, b)])]

getResults2 :: (Eq a, Eq b, Eq c, Eq d, Listable a, Listable b, Listable c, Listable d) => Int -> (a -> b) -> (c -> d) -> ((a -> b) -> (c -> d) -> [Bool]) -> [([Int], Int, Maybe (Mutation a b, Mutation c d), Int, Maybe (Mutation a b, Mutation c d))]

getResults3 :: (Eq a, Eq b, Eq c, Eq d, Eq e, Eq f, Listable a, Listable b, Listable c, Listable d, Listable e, Listable f) => Int -> (a -> b) -> (c -> d) -> (e -> f) -> ((a -> b) -> (c -> d) -> (e -> f) -> [Bool]) -> [([Int], Int, Maybe (Mutation a b, Mutation c d, Mutation e f), Int, Maybe (Mutation a b, Mutation c d, Mutation e f))]

propEvalMap :: Eq a => ((a -> b) -> [Bool]) -> (a -> b) -> [(a, b)] -> [Maybe Bool]

propEvalMap2 :: (Eq a, Eq c) => ((a -> b) -> (c -> d) -> [Bool]) -> (a -> b) -> (c -> d) -> Mutation a b -> Mutation c d -> [Maybe Bool]

propEvalMap3 :: (Eq a, Eq c, Eq e) => ((a -> b) -> (c -> d) -> (e -> f) -> [Bool]) -> (a -> b) -> (c -> d) -> (e -> f) -> Mutation a b -> Mutation c d -> Mutation e f -> [Maybe Bool]

evalMap :: Eq a => ((a -> b) -> [Bool]) -> (a -> b) -> [(a, b)] -> [Bool]

evalMap2 :: (Eq a, Eq c) => ((a -> b) -> (c -> d) -> [Bool]) -> (a -> b) -> (c -> d) -> Mutation a b -> Mutation c d -> [Bool]

evalMap3 :: (Eq a, Eq c, Eq e) => ((a -> b) -> (c -> d) -> (e -> f) -> [Bool]) -> (a -> b) -> (c -> d) -> (e -> f) -> Mutation a b -> Mutation c d -> Mutation e f -> [Bool]

report1With :: (Eq a, Eq b, Show a, Show b, Listable a, Listable b) => CArgs -> Int -> (a -> b) -> ((a -> b) -> [Bool]) -> IO ()

report1 :: (Eq a, Eq b, Show a, Show b, Listable a, Listable b) => Int -> (a -> b) -> ((a -> b) -> [Bool]) -> IO ()

report2With :: (Eq a, Eq b, Eq c, Eq d, Show a, Show b, Show c, Show d, Listable a, Listable b, Listable c, Listable d) => CArgs -> Int -> (a -> b) -> (c -> d) -> ((a -> b) -> (c -> d) -> [Bool]) -> IO ()

report2 :: (Eq a, Eq b, Eq c, Eq d, Show a, Show b, Show c, Show d, Listable a, Listable b, Listable c, Listable d) => Int -> (a -> b) -> (c -> d) -> ((a -> b) -> (c -> d) -> [Bool]) -> IO ()

report3With :: (Eq a, Eq b, Eq c, Eq d, Eq e, Eq f, Show a, Show b, Show c, Show d, Show e, Show f, Listable a, Listable b, Listable c, Listable d, Listable e, Listable f) => CArgs -> Int -> (a -> b) -> (c -> d) -> (e -> f) -> ((a -> b) -> (c -> d) -> (e -> f) -> [Bool]) -> IO ()

report3 :: (Eq a, Eq b, Eq c, Eq d, Eq e, Eq f, Show a, Show b, Show c, Show d, Show e, Show f, Listable a, Listable b, Listable c, Listable d, Listable e, Listable f) => Int -> (a -> b) -> (c -> d) -> (e -> f) -> ((a -> b) -> (c -> d) -> (e -> f) -> [Bool]) -> IO ()

nSurvCT :: Eq pid => [pid] -> (a -> [Maybe Bool]) -> [a] -> [([pid], Int, Maybe a, Int, Maybe a)]

compositions' :: [Maybe Bool] -> [Maybe Bool]