Mutation

Documentation

type Mutation a b = [(a, b)]

canonicalMutation :: Eq b => (a -> b) -> [(a, b)] -> Bool

szMutations :: (Eq b, Listable a, Listable b) => (a -> b) -> [[Mutation a b]]

szMutations2 :: (Eq b, Eq d, Listable a, Listable b, Listable c, Listable d) => (a -> b) -> (c -> d) -> [[(Mutation a b, Mutation c d)]]

szMutations3 :: (Eq b, Eq d, Eq f, Listable a, Listable b, Listable c, Listable d, Listable e, Listable f) => (a -> b) -> (c -> d) -> (e -> f) -> [[(Mutation a b, Mutation c d, Mutation e f)]]

mutations :: (Eq b, Listable a, Listable b) => (a -> b) -> [Mutation a b]

mutations2 :: (Eq b, Eq d, Listable a, Listable b, Listable c, Listable d) => (a -> b) -> (c -> d) -> [(Mutation a b, Mutation c d)]

mutations3 :: (Eq b, Eq d, Eq f, Listable a, Listable b, Listable c, Listable d, Listable e, Listable f) => (a -> b) -> (c -> d) -> (e -> f) -> [(Mutation a b, Mutation c d, Mutation e f)]

makeMutant :: Eq a => (a -> b) -> Mutation a b -> a -> b

makeMutantTraps :: Eq a => (a -> b) -> Mutation a b -> [a -> b]

makeMutantTraps2 :: (Eq a, Eq c) => (a -> b) -> (c -> d) -> Mutation a b -> Mutation c d -> [(a -> b, c -> d)]

makeMutantTraps3 :: (Eq a, Eq c, Eq e) => (a -> b) -> (c -> d) -> (e -> f) -> Mutation a b -> Mutation c d -> Mutation e f -> [(a -> b, c -> d, e -> f)]

makeBindingTraps :: Eq a => Mutation a b -> [Mutation a b]

makeBindingTraps2 :: (Eq a, Eq c) => Mutation a b -> Mutation c d -> [(Mutation a b, Mutation c d)]

makeBindingTraps3 :: (Eq a, Eq c, Eq e) => Mutation a b -> Mutation c d -> Mutation e f -> [(Mutation a b, Mutation c d, Mutation e f)]

showMutant_ :: (Show a, Show b) => String -> String -> Mutation a b -> String